更新

+++ yunjingapp/src/com/yunjing/service/impl/BusinessServiceImpl.java
@@ -675,36 +675,40 @@ public class BusinessServiceImpl implements BusinessService {
 				result.setCode("-1000");
 				result.setDesc("非脉冲防区，本次设置无效");
 			} else {
-				
-				Push push = new Push();
-				push.setZoneNo(zoneNo);
-				push.setMsgId(Utils.getUUID());
-				push.setAddDate(DateUtil.getNowDateTime());
-				push.setPushService("0"); //MQTT推送
-				push.setItype("7"); //消息类型为触网防区参数配置
-				push.setZoneParam(zoneParam);
-				push.setTopic(zoneDb.getDeviceNo()); //topic 为设备编号
-				businessDao.deleteNotPushInvalidMsg(push.getTopic(), push.getItype()); //删除无效批量布撤防推送消息
-				String operatorType = "10"; //触网防区配置
-				
-				PushDto dto = new PushDto();
-				BeanUtils.copyProperties(push, dto);
-				dto.setDeviceNo(zoneDb.getDeviceNo());
-				dto.setTime(System.currentTimeMillis() + "");
-				push.setMsgText(Utils.wrapPushMsg(dto));
-				logger.info("msgText:" + push.getMsgText());
-				businessDao.savePushMsg(push);
-				
-				//添加操作日志
-				OperatorLog log = new OperatorLog();
-				log.setZoneNo(zoneNo);
-				log.setDeviceNo(zoneDb.getDeviceNo());
-				log.setIpAddr(ipAddr);
-				log.setMemberId(userId);
-				log.setMemo("用户进行触网防区配置操作,配置的值为:" + zoneParam);
-				log.setOperatorType(operatorType); //操作类型
-				saveOperatorLog(log);
-				
+				if (zoneDb.getZoneOnline().equals("0")){
+					result.setCode("-1001");
+					result.setDesc("防区未上线不能操作");
+				} else {
+					Push push = new Push();
+					push.setZoneNo(zoneNo);
+					push.setMsgId(Utils.getUUID());
+					push.setAddDate(DateUtil.getNowDateTime());
+					push.setPushService("0"); //MQTT推送
+					push.setItype("7"); //消息类型为触网防区参数配置
+					push.setZoneParam(zoneParam);
+					push.setTopic(zoneDb.getDeviceNo()); //topic 为设备编号
+					businessDao.deleteNotPushInvalidMsg(push.getTopic(), push.getItype()); //删除无效批量布撤防推送消息
+					String operatorType = "10"; //触网防区配置
+					
+					PushDto dto = new PushDto();
+					BeanUtils.copyProperties(push, dto);
+					dto.setDeviceNo(zoneDb.getDeviceNo());
+					dto.setTime(System.currentTimeMillis() + "");
+					push.setMsgText(Utils.wrapPushMsg(dto));
+					logger.info("msgText:" + push.getMsgText());
+					businessDao.savePushMsg(push);
+					
+					//添加操作日志
+					OperatorLog log = new OperatorLog();
+					log.setZoneNo(zoneNo);
+					log.setDeviceNo(zoneDb.getDeviceNo());
+					log.setIpAddr(ipAddr);
+					log.setMemberId(userId);
+					log.setMemo("用户进行触网防区配置操作,配置的值为:" + zoneParam);
+					log.setOperatorType(operatorType); //操作类型
+					saveOperatorLog(log);
+					
+				}
 			}
 		}
 		return result;
